[Madagascar] Hotfix/sli/lock user

+++ coua/src/main/java/com/keepreal/madagascar/coua/model/UserInfo.java
@@ -51,7 +51,7 @@ public class UserInfo {
     @Builder.Default
     private Boolean deleted = false;
     @Builder.Default
-    private Integer lockedUntil = 0;
+    private Long lockedUntil = 0L;
     @Builder.Default
     private Integer state = 0;
     private String unionId;
